Fixed small bug with players list update, different colors for server chat messages, screenshots are stored to different files now, added user status information to http report
git-svn-id: https://ufo2000.svn.sourceforge.net/svnroot/ufo2000/trunk@245 80a73507-52ba-0310-bdc1-b9b4a1856fab

+++ src/video.cpp
@@ -263,8 +263,17 @@ void savescreen()
 	BITMAP *scr = create_bitmap(SCREEN_W, SCREEN_H);
 	blit(screen, scr, 0, 0, 0, 0, SCREEN_W, SCREEN_H);
 
-	save_pcx("snapshot.pcx", scr, gamepal);
-	destroy_bitmap(scr);
+	int num = 1;
+	while (true) {
+		char filename[128];
+		sprintf(filename, "snapshot_%d.pcx", num);
+		if (!exists(filename)) {
+			save_pcx(filename, scr, gamepal);
+			destroy_bitmap(scr);
+			return;
+		}
+		num++;
+	}
 }
 
 #define POLY 0x8408
